Successfully done

+++ SpringDb/src/main/java/com/springmvc/demo/repository/EmployeeRepository.java
@@ -1,11 +1,15 @@
 package com.springmvc.demo.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.springmvc.demo.model.Employee;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long>{
+public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
 
+
+	
 }
